Improved visuals of main menu and some UX features
Altered the look of the main menu by adding a title and using different colors to improve the look. Got strings to display to allow the user to determine what each button does. UX improvements like opening in fullscreen and exiting on escape to make it easier for the user.